+++ pkg/timestream/executor.go
@@ -51,6 +51,10 @@ func ExecuteQuery(ctx context.Context, query models.QueryModel, runner queryRunn
 	if frame.Meta.Custom == nil {
 		frame.Meta.Custom = &models.TimestreamCustomMeta{}
 	}
+	if output.QueryStatus != nil {
+		c := frame.Meta.Custom.(*models.TimestreamCustomMeta)
+		c.Status = output.QueryStatus
+	}
 
 	// Apply the timing info
 	meta := frame.Meta.Custom.(*models.TimestreamCustomMeta)