+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 ## v1.1.0
 
 - Updated authentication to match builtin cloudwatch authentication
+- Include query status in metadata
 
 ## v1.0.0